Iteracion de documentacion
Se ihcieron mejoras separando la funcionalidad de lo reportes generados por el sistema en sus propias clases

Ademas ahora se pueden añadir nuevos documentos al sistema
