+++ ProfessionalPracticesSystem/GUI-WPF/Pages/Practitioner/Documentation.xaml
@@ -25,6 +25,16 @@
                         </StackPanel>
 
                         <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,20,0,0">
+                            <TextBlock Text="Oficio de aceptación" FontSize="18" TextWrapping="Wrap" VerticalAlignment="Center" Margin="0,0,50,0" Width="200"/>
+                            <Button Margin="20,0,20,0" Width="135" Content="Añadir" Click="AddAceptanceLetter"/>
+                        </StackPanel>
+
+                        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,20,0,0">
+                            <TextBlock Text="Oficio de asignación" FontSize="18" TextWrapping="Wrap" VerticalAlignment="Center" Margin="0,0,50,0" Width="200"/>
+                            <Button Margin="20,0,20,0" Width="135" Content="Añadir" Click="AddAssigmentLetter"/>
+                        </StackPanel>
+
+                        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,20,0,0">
                             <TextBlock Text="Reporte parcial" FontSize="18" TextWrapping="Wrap" VerticalAlignment="Center" Margin="0,0,50,0" Width="200"/>
                             <Button Margin="0,0,20,0" Click="GeneratePartialReport">Generar</Button>
                             <Button Click="AddPartialReport">Añadir</Button>
