+++ ProfessionalPracticesSystem/GUI-WPF/Pages/Professor/MensualReports.xaml.cs
@@ -1,0 +1,30 @@
+﻿using DataAccess.Implementation;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace GUI_WPF.Pages.Professor
+{
+    /// <summary>
+    /// Lógica de interacción para MensualReports.xaml
+    /// </summary>
+    public partial class MensualReports : Page
+    {
+        private CollectionView nombre;
+        public MensualReports()
+        {
+            InitializeComponent();
+        }
+    }
+}
